Refactor Entity and World classes: replace `int_fast8_t` with `int` for position handling and adjust world dimensions

+++ include/entity.hpp
@@ -1,7 +1,6 @@
 #pragma once
 
 #include "stats.hpp"
-#include <stdint.h>
 #include <string>
 #include <string_view>
 #include <utility>
@@ -18,8 +17,8 @@ class Entity {
     virtual void take_dmg(int val) noexcept;
     virtual void heal(int val) noexcept;
     virtual bool is_alive() const noexcept;
-    int_fast8_t get_x_pos() const noexcept;
-    int_fast8_t get_y_pos() const noexcept;
+    int get_x_pos() const noexcept;
+    int get_y_pos() const noexcept;
 
     virtual ~Entity() = default;
 
@@ -27,12 +26,12 @@ class Entity {
     void set_attack(int value) noexcept;
     void set_defence(int value) noexcept;
     void set_hp(int value) noexcept;
-    void set_position(int_fast8_t new_x_pos, int_fast8_t new_y_pos) noexcept;
+    void set_position(int new_x_pos, int new_y_pos) noexcept;
 
   private:
     std::string name_{};
     Stats stats_{};
     int hp_{};
-    int_fast8_t x_pos{};
-    int_fast8_t y_pos{};
+    int x_pos{};
+    int y_pos{};
 };
